Make LeavePrivateChannelModal generic
