+++ components/channel_header_dropdown/menu_items/leave_channel/leave_channel.tsx
@@ -9,7 +9,7 @@ import {Constants, ModalIdentifiers} from 'utils/constants';
 import {localizeMessage} from 'utils/utils';
 
 import Menu from 'components/widgets/menu/menu';
-import LeavePrivateChannelModal from 'components/leave_private_channel_modal';
+import LeaveChannelModal from 'components/leave_channel_modal';
 
 import type {PropsFromRedux} from './index';
 
@@ -56,7 +56,7 @@ export default class LeaveChannel extends React.PureComponent<Props> {
         if (channel.type === Constants.PRIVATE_CHANNEL) {
             openModal({
                 modalId: ModalIdentifiers.LEAVE_PRIVATE_CHANNEL_MODAL,
-                dialogType: LeavePrivateChannelModal,
+                dialogType: LeaveChannelModal,
                 dialogProps: {
                     channel,
                 },
