+++ i18n/en.json
@@ -2799,7 +2799,6 @@
   "channel_header.groupMessageHasGuests": "This group message has guests",
   "channel_header.lastActive": "Active {timestamp}",
   "channel_header.leave": "Leave Channel",
-  "channel_header.leaveConfirm": "Are you sure you want to leave the channel?",
   "channel_header.manageMembers": "Manage Members",
   "channel_header.menuAriaLabel": "Channel Menu",
   "channel_header.mute": "Mute Channel",
@@ -2859,7 +2858,6 @@
   "channel_members_dropdown.make_channel_members": "Make Channel Members",
   "channel_members_dropdown.menuAriaLabel": "Change the role of channel member",
   "channel_members_dropdown.remove_from_channel": "Remove from Channel",
-  "channel_members_dropdown.remove_from_channelConfirm": "Are you sure you want to remove user from the channel?",
   "channel_members_dropdown.shared_member": "Shared Member",
   "channel_members_modal.addNew": " Add Members",
   "channel_members_modal.members": " Members",
@@ -3874,6 +3872,8 @@
   "leave_private_channel_modal.leave": "Yes, leave channel",
   "leave_private_channel_modal.message": "Are you sure you wish to leave the private channel {channel}? You must be re-invited in order to re-join this channel in the future.",
   "leave_private_channel_modal.title": "Leave Private Channel {channel}",
+  "leave_public_channel_modal.message": "Are you sure you wish to leave the channel {channel}? You can re-join this channel in the future if you change your mind.",
+  "leave_public_channel_modal.title": "Leave Channel {channel}",
   "leave_team_modal_guest_only_private.desc": "**You will be removed from {num_of_private_channels} private { num_of_private_channels,plural,one {channel} other {channels}} on this team.** You won't be able to rejoin it without an invitation from another team member. Are you sure?",
   "leave_team_modal_guest_only_public.desc": "**You will be removed from {num_of_public_channels} public { num_of_public_channels,plural,one {channel} other {channels}} on this team.** You won't be able to rejoin it without an invitation from another team member. Are you sure?",
   "leave_team_modal_guest.desc": "**You will be removed from {num_of_public_channels} public { num_of_public_channels,plural,one {channel} other {channels}} and {num_of_private_channels} private { num_of_private_channels,plural,one {channel} other {channels}} on this team.** You won't be able to rejoin it without an invitation from another team member. Are you sure?",
